Minor gcc warn fixes
<==== Mercurial data ====>
Branch: master

+++ common/src/plugins/resources/archive/archive_stream_reader.cpp
@@ -755,7 +755,7 @@ begin_label:
         m_playbackMaskSync.unlock();
 
         qint64 maxTime = m_delegate->endTime();
-        if (newTime == DATETIME_NOW || newTime == -1 || (maxTime != AV_NOPTS_VALUE && newTime > maxTime)) {
+        if (newTime == DATETIME_NOW || newTime == -1 || (maxTime != (qint64)AV_NOPTS_VALUE && newTime > maxTime)) {
             //internalJumpTo(qMax(0ll, newTime)); // seek to end or BOF.
             m_outOfPlaybackMask = true;
             return createEmptyPacket(reverseMode); // EOF reached
